Add delete confirmation, Fix bugs

+++ AddressBookUI/ViewModels/DisplayPersonViewModel.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using AddressBookLibrary.Models;
 using Caliburn.Micro;
 
@@ -22,7 +23,14 @@ namespace AddressBookUI.ViewModels
 
         public void Delete()
         {
-            shell.DeleteSelectedPerson();
+            if (SelectedPerson != null)
+            {
+                MessageBoxResult result = MessageBox.Show($"Are you sure you want to delete {SelectedPerson.FullName}?", "Delete Confirmation", MessageBoxButton.YesNo);
+                if (result == MessageBoxResult.Yes)
+                {
+                    shell.DeleteSelectedPerson();
+                }
+            }     
         }
 
         public void Edit()
